@@ -16,12 +16,13 @@
   </modules>
 
   <properties>
-    <java.version>14</java.version>
-    <dropwizard.version>2.1.4</dropwizard.version>
-    <pact.version>4.0.10</pact.version>
+    <java.version>19</java.version>
+    <dropwizard.version>4.0.2</dropwizard.version>
+    <pact.version>4.4.9</pact.version>
     <kotlin.version>1.9.10</kotlin.version>
-    <jersey.version>2.37</jersey.version>
+    <jersey.version>3.0.11</jersey.version>
     <netty.version>4.1.100.Final</netty.version>
+    <junit.version>5.10.0</junit.version>
     <error_prone.version>2.22.0</error_prone.version>
     <mvn-test-plugins.version>3.1.2</mvn-test-plugins.version>
     <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
