@@ -1,32 +1,37 @@
 package specialmembershipservice.it.pacts.verifications;
 
-import static com.github.restdriver.clientdriver.RestClientDriver.giveResponse;
+import static com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder.responseDefinition;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.common.ContentTypes.APPLICATION_JSON;
+import static com.github.tomakehurst.wiremock.common.ContentTypes.CONTENT_TYPE;
+import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
 import static java.util.Collections.singletonList;
-import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
+import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
-import static org.junit.Assert.assertThat;
 import static specialmembershipservice.it.pacts.PactConstants.PACT_BROKER_PORT;
 import static specialmembershipservice.it.pacts.PactConstants.PACT_BROKER_URL;
 import static specialmembershipservice.it.pacts.PactConstants.SPECIAL_MEMBERSHIP_SERVICE;
 
 import au.com.dius.pact.provider.PactVerifyProvider;
-import au.com.dius.pact.provider.junit.IgnoreNoPactsToVerify;
-import au.com.dius.pact.provider.junit.MessagePactRunner;
-import au.com.dius.pact.provider.junit.Provider;
-import au.com.dius.pact.provider.junit.State;
-import au.com.dius.pact.provider.junit.loader.PactBroker;
-import au.com.dius.pact.provider.junit.loader.PactBrokerAuth;
-import au.com.dius.pact.provider.junit.target.AmqpTarget;
-import au.com.dius.pact.provider.junit.target.Target;
-import au.com.dius.pact.provider.junit.target.TestTarget;
+import au.com.dius.pact.provider.junit5.MessageTestTarget;
+import au.com.dius.pact.provider.junit5.PactVerificationContext;
+import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
+import au.com.dius.pact.provider.junitsupport.IgnoreNoPactsToVerify;
+import au.com.dius.pact.provider.junitsupport.Provider;
+import au.com.dius.pact.provider.junitsupport.State;
+import au.com.dius.pact.provider.junitsupport.loader.PactBroker;
+import au.com.dius.pact.provider.junitsupport.loader.PactBrokerAuth;
+import com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder;
+import com.github.tomakehurst.wiremock.junit5.WireMockExtension;
+import jakarta.ws.rs.core.Response;
 import java.util.Map;
-import javax.ws.rs.core.Response;
-import org.junit.Rule;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Order;
+import org.junit.jupiter.api.TestTemplate;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.jupiter.api.extension.RegisterExtension;
 import specialmembershipservice.it.IntegrationTestBase;
-import specialmembershipservice.it.creditscoreservice.CreditScoreServiceRule;
 
-@RunWith(MessagePactRunner.class)
 @Provider(SPECIAL_MEMBERSHIP_SERVICE)
 @IgnoreNoPactsToVerify
 @PactBroker(
@@ -38,19 +43,28 @@ import specialmembershipservice.it.creditscoreservice.CreditScoreServiceRule;
 )
 public class MessagePactVerifications extends IntegrationTestBase {
 
-  @Rule
-  public final CreditScoreServiceRule creditScoreServiceRule = new CreditScoreServiceRule(
-      CREDIT_SCORE_SERVICE_PORT);
+  @RegisterExtension
+  @Order(0)
+  private static WireMockExtension CREDIT_SCORE_SERVICE_RULE = WireMockExtension.newInstance()
+      .options(wireMockConfig().port(CREDIT_SCORE_SERVICE_PORT))
+      .build();
 
-  @TestTarget
-  public final Target target = new AmqpTarget(
-      singletonList(this.getClass().getPackage().getName()));
+  @TestTemplate
+  @ExtendWith(PactVerificationInvocationContextProvider.class)
+  public void pactVerificationTestTemplate(PactVerificationContext context) {
+    context.verifyInteraction();
+  }
+
+  @BeforeEach
+  public void before(PactVerificationContext context) {
+    context.setTarget(new MessageTestTarget(singletonList(this.getClass().getPackage().getName())));
+  }
 
   @State("Tony Stark became a new member")
   public void tonyStarkBecameANewMember() {
     String email = "tony.stark@example.com";
-    creditScoreServiceRule.setCreditResponse(email,
-        giveResponse(creditScoreDto(850), APPLICATION_JSON));
+    setCreditResponse(email, responseDefinition().withBody(creditScoreDto(850))
+        .withHeader(CONTENT_TYPE, APPLICATION_JSON));
     Map<String, Object> specialMembershipDto = specialMembershipDto(email);
     Response response = resourcesClient.postSpecialMembership(specialMembershipDto);
     response.close();
@@ -60,5 +74,9 @@ public class MessagePactVerifications extends IntegrationTestBase {
   @PactVerifyProvider("An event notifying Tony Stark's new membership")
   public String verifyTonyStarksNewMembershipEvent() throws Exception {
     return readPublishedMessage();
+  }
+
+  protected void setCreditResponse(String email, ResponseDefinitionBuilder response) {
+    CREDIT_SCORE_SERVICE_RULE.stubFor(get("/credit-scores/" + email).willReturn(response));
   }
 }
