@@ -1,26 +1,26 @@
 package welcomememberemailservice.it.pacts;
 
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static welcomememberemailservice.it.pacts.PactConstants.SPECIAL_MEMBERSHIP_SERVICE;
 import static welcomememberemailservice.it.pacts.PactConstants.WELCOME_MEMBER_EMAIL_SERVICE;
 
 import au.com.dius.pact.consumer.MessagePactBuilder;
 import au.com.dius.pact.consumer.dsl.PactDslJsonBody;
-import au.com.dius.pact.consumer.junit.MessagePactProviderRule;
-import au.com.dius.pact.consumer.junit.PactVerification;
+import au.com.dius.pact.consumer.junit5.PactConsumerTestExt;
+import au.com.dius.pact.consumer.junit5.PactTestFor;
+import au.com.dius.pact.consumer.junit5.ProviderType;
+import au.com.dius.pact.core.model.PactSpecVersion;
 import au.com.dius.pact.core.model.annotations.Pact;
+import au.com.dius.pact.core.model.messaging.Message;
 import au.com.dius.pact.core.model.messaging.MessagePact;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
-import org.junit.Rule;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import welcomememberemailservice.it.IntegrationTestBase;
 
+@ExtendWith(PactConsumerTestExt.class)
 public class MemberSignedUpEventPactContracts extends IntegrationTestBase {
-
-  @Rule
-  public final MessagePactProviderRule specialMembershipServiceRule =
-      new MessagePactProviderRule(SPECIAL_MEMBERSHIP_SERVICE, this);
 
   @Pact(consumer = WELCOME_MEMBER_EMAIL_SERVICE, provider = SPECIAL_MEMBERSHIP_SERVICE)
   public MessagePact newMemberTonyStark(MessagePactBuilder builder) {
@@ -37,10 +37,11 @@ public class MemberSignedUpEventPactContracts extends IntegrationTestBase {
         .toPact();
   }
 
-  @PactVerification(value = SPECIAL_MEMBERSHIP_SERVICE, fragment = "newMemberTonyStark")
   @Test
-  public void sendWelcomeEmailToTonyStark() throws Exception {
-    String memberSignedUpEvent = new String(specialMembershipServiceRule.getMessage(), UTF_8);
+  @PactTestFor(pactMethod = "newMemberTonyStark",
+      providerType = ProviderType.ASYNCH, pactVersion = PactSpecVersion.V3)
+  public void sendWelcomeEmailToTonyStark(List<Message> messages) {
+    String memberSignedUpEvent = messages.get(0).contentsAsString();
     publishMembershipMessageAndWaitToBeConsumed(memberSignedUpEvent);
     assertAnEmailWasSent();
   }
